backend/requirements.txt
# ============================================
# MILANO - Backend Dependencies (OPTIMIZADO)
# ============================================

# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# Pydantic - FIX: Actualizado a v2 para mejor performance
pydantic==2.5.0
pydantic-settings==2.1.0

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1  # FIX: Actualizado

# Utilidades
python-dotenv==1.0.0

# FIX: Añadidos para logging y monitoring
structlog==23.2.0

# FIX: Añadido para validaciones adicionales
email-validator==2.1.0